Make Game helper
Put the collision helper inside the game helper as a member.
